Source files for Isosurface project
CMakeLists.txt: A type of make file to be used with the CMake program.
Isosurface.cxx: The main source code for the project.
TriangleList.h: A class file with inline functions to store information
from the main source file.
Isosurface.vtk: Data to be used by the final executable. This file MUST
be moved into the same folder as the executable.